+++ src/bucket.rs
@@ -1,3 +1,5 @@
+use std::convert::From;
+
 pub const FINGERPRINT_SIZE: usize = 1;
 pub const BUCKET_SIZE: usize = 4;
 const EMPTY_FINGERPRINT_DATA: [u8; FINGERPRINT_SIZE] = [100; FINGERPRINT_SIZE];
@@ -23,12 +25,19 @@ impl Fingerprint {
 
     /// Returns the empty Fingerprint.
     pub fn empty() -> Fingerprint {
-        Fingerprint { data: EMPTY_FINGERPRINT_DATA }
+        Fingerprint {
+            data: EMPTY_FINGERPRINT_DATA,
+        }
     }
 
     /// Checks if this is the empty Fingerprint.
     pub fn is_empty(&self) -> bool {
         self.data == EMPTY_FINGERPRINT_DATA
+    }
+
+    /// Sets the fingerprint value to a previously exported one via an in-memory copy.
+    fn slice_copy(&mut self, fingerprint: &[u8]) {
+        self.data.copy_from_slice(fingerprint);
     }
 }
 
@@ -41,11 +50,13 @@ pub struct Bucket {
 impl Bucket {
     /// Creates a new bucket with a pre-allocated buffer.
     pub fn new() -> Bucket {
-        Bucket { buffer: [Fingerprint::empty(); BUCKET_SIZE] }
+        Bucket {
+            buffer: [Fingerprint::empty(); BUCKET_SIZE],
+        }
     }
 
-    /// Inserts the fingerprint into the buffer if the buffer is not full. This
-    /// operation is O(1).
+    /// Inserts the fingerprint into the buffer if the buffer is not full.
+    /// This operation is O(1).
     pub fn insert(&mut self, fp: Fingerprint) -> bool {
         for entry in &mut self.buffer {
             if entry.is_empty() {
@@ -70,5 +81,25 @@ impl Bucket {
     /// Returns the index of the given fingerprint, if its found. O(1)
     pub fn get_fingerprint_index(&self, fp: Fingerprint) -> Option<usize> {
         self.buffer.iter().position(|e| *e == fp)
+    }
+
+    /// Returns all current fingerprint data of the current buffer for storage.
+    pub fn get_fingerprint_data(&self) -> Vec<u8> {
+        self.buffer
+            .iter()
+            .flat_map(|f| f.data.into_iter())
+            .cloned()
+            .collect()
+    }
+}
+
+impl<'a> From<&'a [u8]> for Bucket {
+    /// Constructs a buffer of fingerprints from a set of previously exported fingerprints.
+    fn from(fingerprints: &'a [u8]) -> Self {
+        let mut buffer = [Fingerprint::empty(); BUCKET_SIZE];
+        for (idx, value) in fingerprints.chunks(FINGERPRINT_SIZE).enumerate() {
+            buffer[idx].slice_copy(value);
+        }
+        Bucket { buffer: buffer }
     }
 }